+++ src/main/scala/xitrum/RequestVar.scala
@@ -1,5 +1,5 @@
 package xitrum
 
-class RequestVar[+A] extends OptVar[A] {
+class RequestVar[+A](implicit m: Manifest[A]) extends OptVar[A] {
   def getAll(implicit action: Action) = action.at
 }